Fixes #34698 - add dropdown and move file
